Docs : Modal Component에서 Rental에 관련한 JSX 분리
